requirements.txt
PyYAML
cookiecutter
